Move Mock Web Service to its own repo
