+++ Package.swift
@@ -18,13 +18,11 @@ let package = Package(
         .iOS(.v16)
     ],
     products: [
-        .library(name: "SpeziFHIR", targets: ["SpeziFHIR"]),
-        .library(name: "SpeziMockWebService", targets: ["SpeziMockWebService"])
+        .library(name: "SpeziFHIR", targets: ["SpeziFHIR"])
     ],
     dependencies: [
         .package(url: "https://github.com/apple/FHIRModels", .upToNextMinor(from: "0.5.0")),
-        .package(url: "https://github.com/StanfordSpezi/Spezi", .upToNextMinor(from: "0.7.0")),
-        .package(url: "https://github.com/StanfordSpezi/SpeziViews", .upToNextMinor(from: "0.4.0"))
+        .package(url: "https://github.com/StanfordSpezi/Spezi", .upToNextMinor(from: "0.7.0"))
     ],
     targets: [
         .target(
@@ -38,17 +36,6 @@ let package = Package(
             name: "SpeziFHIRTests",
             dependencies: [
                 .target(name: "SpeziFHIR")
-            ]
-        ),
-        .target(
-            name: "SpeziMockWebService",
-            dependencies: [
-                .target(name: "SpeziFHIR"),
-                .product(name: "Spezi", package: "Spezi"),
-                .product(name: "SpeziViews", package: "SpeziViews")
-            ],
-            resources: [
-                .process("Resources")
             ]
         )
     ]
