--- Tests/UITests/TestApp/Info.plist.license
@@ -1,5 +1,0 @@
-This source file is part of the Stanford Spezi open-source project
-
-SPDX-FileCopyrightText: 2022 Stanford University and the project authors (see CONTRIBUTORS.md)
-
-SPDX-License-Identifier: MIT
