+++ Tests/UITests/TestAppUITests/FHIRMockDataStorageProviderTests.swift
@@ -7,8 +7,6 @@
 //
 
 import XCTest
-import XCTestExtensions
-import XCTHealthKit
 
 
 final class FHIRMockWebServiceTests: XCTestCase {
@@ -16,22 +14,6 @@ final class FHIRMockWebServiceTests: XCTestCase {
         let app = XCUIApplication()
         app.launch()
         
-        app.buttons["FHIRMockWebService"].tap()
-        
-        XCTAssert(app.buttons["Mock Requests"].waitForExistence(timeout: 2))
-        app.buttons["Mock Requests"].tap()
-        
-        XCTAssert(app.images["Previous Page"].waitForExistence(timeout: 2))
-        app.images["Previous Page"].tap()
-        
-        XCTAssert(app.staticTexts["/Test/"].waitForExistence(timeout: 2))
-        XCTAssert(app.staticTexts["{\"test\": \"test\"}"].waitForExistence(timeout: 2))
-        
-        app.navigationBars.buttons["Mock Upload"].tap()
-        
-        XCTAssert(app.images["Trash"].waitForExistence(timeout: 2))
-        app.images["Trash"].tap()
-        
-        XCTAssert(app.staticTexts["/TestRemoval/"].waitForExistence(timeout: 2))
+        XCTAssert(app.staticTexts["FHIR"].waitForExistence(timeout: 2))
     }
 }
